fix(ugc): sync actions for deep-linked images
